+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added `DefaultReserver` which works just like `OrderedReserver` with one exception: it does not sort the queues
 
 ## [3.0.0] - 2018-10-19
 ### Changed